Main Scene Finished
The Main scene is finished and fully functional. Three different shapes with different names and colors are in a row. They can be clicked on to have them tell the player their name and color. The code used to do this illustrates the the four pillars of OOP; inheritance, polymorphism, encapsulation, and abstraction.